everything but vehicles and requests
