+++ pyrai/dispatcher/methods/fleet/make_vehicle_offline.py
@@ -1,16 +1,17 @@
 import requests
 import datetime
 import json
-from pyrai.dispatcher.structures import Endpoints, StatusResponse, StatusError
+from pyrai.dispatcher.structures.endpoints import Endpoints
+from pyrai.dispatcher.structures.status_response import StatusResponse
+from pyrai.dispatcher.structures.status_error import StatusError
 
-def make_vehicle_online(self, vid, location, capacity):
+def make_vehicle_offline(self, vid, location):
     """
-    Attempts to make vehicle online.
+    Attempts to take vehicle offline.
 
     Args:
         vid (int): The vehicle ID.
         location (Location): The vehicle location.
-        capacity (int): The vehicle capacity.
 
     Raises:
         StatusError: If unsuccessful.
@@ -18,20 +19,18 @@ def make_vehicle_online(self, vid, location, capacity):
     Returns:
         StatusReponse: If successful.
     """
-
     if datetime.datetime.now() > self.end_time:
         self.end_time = datetime.datetime.now()
 
-    url = self.build_url(Endpoints.MAKE_VEHICLE_ONLINE)
+    url = self.build_url(Endpoints.MAKE_VEHICLE_OFFLINE)
     payload = {
-        "location": location.todict(),
-        "id": vid,
-        'capacity': capacity,
+        'location': location.todict(),
+        'id': vid,
         'user_key': self.user_key.todict()
     }
-    r = requests.post(url, data=json.dumps(payload))
+    r = requests.post(url, data = json.dumps(payload))
     resp = r.json()
-
+    
     if r.status_code == 200:
         return StatusResponse(resp=resp)
     else:
