+++ pyrai/dispatcher/methods/fleet/remove_vehicle.py
@@ -1,7 +1,8 @@
 import datetime
 import requests
 import json
-from pyrai.dispatcher.structures import Endpoints, StatusError
+from pyrai.dispatcher.structures.endpoints import Endpoints
+from pyrai.dispatcher.structures.status_error import StatusError
 
 def remove_vehicle(self, vid, location):
     """
