+++ pyrai/helpers/__init__.py
@@ -1,2 +1,3 @@
-from .helpers import to_rfc3339
-__all__ = ["to_rfc3339"]
+from .helpers import to_rfc3339, build_url
+__all__ = ["to_rfc3339",
+"build_url"]
